refactored Component names and folder structure - added achievement component
