+++ frontend/src/App.vue
@@ -3,8 +3,8 @@ import { storeToRefs } from 'pinia';
 import { RouterView } from 'vue-router';
 import { useUserStore } from './store/user';
 import { useErrorStore } from './store/error';
-import ModalComponent from './components/ModalComponent.vue';
-import NavbarComponent from './components/NavbarComponent.vue';
+import ModalComponent from './components/globalUse/ModalComponent.vue';
+import NavbarComponent from './components/globalUse/NavbarComponent.vue';
 
 const userStore = useUserStore();
 const errorStore = useErrorStore();
