+++ frontend/src/components/chat/channels/ChildChannelComponent.vue
