+++ frontend/src/components/chat/channels/ModalChannelComponent.vue
