+++ frontend/src/components/chat/chat/ParentChatComponent.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref, onUnmounted } from 'vue';
 import type { Ref } from 'vue';
-import { socket } from '../plugin/socket';
+import { socket } from '../../../plugin/socket';
 
 const props = defineProps<{
   chatName: string;
