+++ frontend/src/components/chat/options/ParentOptionsComponent.vue
