+++ frontend/src/components/chat/peoples/ChildPeopleComponent.vue
