+++ frontend/src/components/game/PongComponent.vue
