+++ frontend/src/components/globalUse/NavbarComponent.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { RouterLink } from 'vue-router';
-import { useUserStore } from '../store/user';
+import { useUserStore } from '../../store/user';
 const userStore = useUserStore();
 </script>
