+++ frontend/src/components/landingview/CategoryComponent.vue
