+++ frontend/src/components/landingview/CategorySearchComponent.vue
