+++ frontend/src/components/leaderboard/ChildMedalComponent.vue
