+++ frontend/src/components/leaderboard/ChildPlayerComponent.vue
