+++ frontend/src/components/profile/aboutMe/AboutMeComponent.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useUserStore } from '../store/user';
+import { useUserStore } from '../../../store/user';
 
 const userStore = useUserStore();
 </script>
