+++ frontend/src/components/profile/profile/ProfileComponent.vue
