+++ frontend/src/components/skinShop/PayPalButtonComponent.vue
