+++ frontend/src/components/stream/ChildStreamComponent.vue
