+++ frontend/src/components/stream/SearchResultComponent.vue
