+++ frontend/src/views/PongView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import GameComponent from '../components/PongComponent.vue';
+import GameComponent from '../components/game/PongComponent.vue';
 </script>
 
 <template>
